feat(embed): add embedHost, embedPath, and embedPort to options. Centralize string-building.
