+++ src/content-handlers/iiif/extensions/uv-av-extension/config/config.json
@@ -285,6 +285,7 @@
       "options": {
         "embedEnabled": true,
         "shareEnabled": true,
+        "embedPath": "/uv.html",
         "embedTemplate": "<iframe src=\"{0}#?{1}\" width=\"{2}\" height=\"{3}\" allowfullscreen frameborder=\"0\" title=\"{4}\"></iframe>",
         "instructionsEnabled": false,
         "shareFrameEnabled": true,
