+++ src/content-handlers/iiif/extensions/uv-mediaelement-extension/config/config.json
@@ -65,6 +65,7 @@
       "options": {
         "embedEnabled": true,
         "shareEnabled": true,
+        "embedPath": "/uv.html",
         "embedTemplate": "<iframe src=\"{0}#?{1}\" width=\"{2}\" height=\"{3}\" allowfullscreen frameborder=\"0\" title=\"{4}\"></iframe>",
         "instructionsEnabled": false,
         "shareFrameEnabled": true,
