+++ src/content-handlers/iiif/extensions/uv-openseadragon-extension/Extension.ts
@@ -1392,18 +1392,7 @@ export default class OpenSeadragonExtension extends BaseExtension<Config> {
       r: rotation.toString(),
     });
 
-    const appUri: string = this.getAppUri();
-    const title: string = this.helper.getLabel() || "";
-    const script: string = Strings.format(
-      template,
-      appUri,
-      hashParams.toString(),
-      width.toString(),
-      height.toString(),
-      title
-    );
-
-    return script;
+    return super.buildEmbedScript(template, width, height, hashParams);
   }
 
   isSearchEnabled(): boolean {
