+++ src/content-handlers/iiif/extensions/uv-openseadragon-extension/config/config.json
@@ -447,6 +447,7 @@
     "shareDialogue": {
       "options": {
         "embedEnabled": true,
+        "embedPath": "/uv.html",
         "embedTemplate": "<iframe src=\"{0}#?{1}\" width=\"{2}\" height=\"{3}\" allowfullscreen frameborder=\"0\" title=\"{4}\"></iframe>",
         "instructionsEnabled": false,
         "shareEnabled": true,
