+++ src/content-handlers/iiif/extensions/uv-pdf-extension/config/config.json
@@ -291,6 +291,7 @@
       "options": {
         "embedEnabled": true,
         "shareEnabled": true,
+        "embedPath": "/uv.html",
         "embedTemplate": "<iframe src=\"{0}#?{1}\" width=\"{2}\" height=\"{3}\" allowfullscreen frameborder=\"0\" title=\"{4}\"></iframe>",
         "instructionsEnabled": false,
         "shareFrameEnabled": true,
