+++ src/content-handlers/iiif/modules/uv-shared-module/BaseExtension.ts
@@ -764,11 +764,6 @@ export class BaseExtension<T extends BaseConfig> implements IExtension {
     return parts.host;
   }
 
-  getAppUri(): string {
-    const port = window.location.port ? `:${window.location.port}` : "";
-    return `${window.location.protocol}//${window.location.hostname}${port}/uv.html`;
-  }
-
   getSettings(): ISettings {
     if (Bools.getBool(this.data.config!.options.saveUserSettings, false)) {
       const settings: any = Storage.get("uv.settings", StorageType.LOCAL);
@@ -818,6 +813,32 @@ export class BaseExtension<T extends BaseConfig> implements IExtension {
       title: title,
       image: thumbnail,
     };
+  }
+
+  getAppUri(): string {
+    const options = this.data.config!.modules.shareDialogue.options;
+
+    const host = options?.embedHost ?? `${window.location.protocol}//${window.location.hostname}`;
+    const port = options?.embedPort ?? window.location.port;
+    const path = options?.embedPath ?? "/uv.html";
+
+    return `${host}${port ? `:${port}` : ""}${path}`;
+  }
+
+  buildEmbedScript(template: string, width: number, height: number, hashParams: URLSearchParams): string {
+    const appUri: string = this.getAppUri();
+    const title: string = this.helper.getLabel() ?? "";
+
+    const script: string = Strings.format(
+      template,
+      appUri,
+      hashParams.toString(),
+      width.toString(),
+      height.toString(),
+      title
+    );
+
+    return script;
   }
 
   public getPagedIndices(
