+++ src/content-handlers/iiif/modules/uv-shared-module/css/scaffolding.less
@@ -48,6 +48,13 @@
     font-family: inherit;
     font-size: inherit;
     line-height: inherit;
+    color: inherit;
+  }
+
+  select,
+  textarea {
+    color: black;
+    background-color: white;
   }
 
   // Button normalization
